requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
bcrypt
python-multipart
python-jose[cryptography]
google-generativeai
python-dotenv
